requirements.txt
# Web Framework
streamlit==1.29.0

# Database
supabase>=2.7.0
httpx>=0.25.0
websockets>=13.0

# Kriptografi
cryptography==41.0.7
bcrypt==4.1.2
pycryptodome==3.19.1

# Image Processing (untuk steganografi)
Pillow>=10.0.0

# Utilities
python-dotenv==1.0.0
